Modified threaded workflow to run single task in main thread of only 1 task is available on join changed the way threadworkload works, you now just pass a void(void) lambda fixed debug assertion error (I think) in bullet system

+++ src/components/weapon/burst/random_shot_weapon_component.cc
@@ -14,8 +14,8 @@ void RandomShotWeaponComponent::Fire(sf::Vector2f position, BulletConfig& config
 {
 	for (float i = 0; i < numBullets; i++)
 	{
-		auto theta = AngleConversion::ToRadians(randSource->Generate(0, 360));
-		auto speed = config.speed * randSource->Generate(50, 250) / 100;
+		auto theta = AngleConversion::ToRadians((float)randSource->Generate(0, 360));
+		auto speed = config.speed * (float)randSource->Generate(50, 250) / 100;
 		sf::Vector2f arcVelocity(std::cos(theta), std::sin(theta));
 		auto traj = BulletTrajectory(position, arcVelocity, speed);
 		this->bulletSystem->FireBullet(factory, traj, config);
